#35 변경사항 적용

+++ src/main/groovy/com/example/refmonolithicserver/ingredient/service/InventoryWriteService.java
@@ -1,6 +1,10 @@
 package com.example.refmonolithicserver.ingredient.service;
 
+import com.example.refmonolithicserver.common.exception.BusinessException;
+import com.example.refmonolithicserver.common.exception.ErrorCode;
+import com.example.refmonolithicserver.common.exception.ExceptionMessage;
 import com.example.refmonolithicserver.ingredient.dao.IngredientRepository;
+import com.example.refmonolithicserver.recipe.dao.RecipeRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,6 +17,7 @@ import static com.example.refmonolithicserver.ingredient.dto.IngredientDto.Ingre
 public class InventoryWriteService {
     
     private final IngredientRepository ingredientRepository;
+    private final RecipeRepository recipeRepository;
 
     public Object addItem(IngredientRequestDto dto, String username) {
         return ingredientRepository.save(dto.toEntity(username));
@@ -23,6 +28,8 @@ public class InventoryWriteService {
     }
 
     public Long removeItem(Long id) {
+        if (recipeRepository.existsByIngredientId(id))
+            throw new BusinessException(ErrorCode.INVALID_PARAMETER, ExceptionMessage.RECIPE_ALREADY_EXIST);
         ingredientRepository.deleteById(id);
         return id;
     }
